SOCINT-291 match changes for separate firestore provider
