+++ pom.xml
@@ -11,7 +11,7 @@
   <parent>
     <groupId>ons.sdc.int.common</groupId>
     <artifactId>common</artifactId>
-    <version>1.0.66</version>
+    <version>1.0.69-SNAPSHOT</version>
   </parent>
 
   <properties>
@@ -115,7 +115,7 @@
     <dependency>
       <groupId>ons.sdc.int.common</groupId>
       <artifactId>cucumber-common</artifactId>
-      <version>0.0.3</version>
+      <version>0.0.4-SNAPSHOT</version>
       <scope>test</scope>
     </dependency>
