+++ src/test/java/uk/gov/ons/ctp/integration/rhcucumber/glue/CucumberSpringConfiguration.java
@@ -4,7 +4,8 @@ import io.cucumber.spring.CucumberContextConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.SpringBootTest;
 import uk.gov.ons.ctp.common.cloud.FirestoreDataStore;
-import uk.gov.ons.ctp.common.cloud.TestCloudDataStore;
+import uk.gov.ons.ctp.common.cloud.FirestoreProviderImpl;
+import uk.gov.ons.ctp.common.firestore.TestCloudDataStore;
 import uk.gov.ons.ctp.common.util.WebDriverFactory;
 import uk.gov.ons.ctp.integration.rhcucumber.repository.RespondentDataRepository;
 import uk.gov.ons.ctp.integration.rhcucumber.selenium.pages.Pages;
@@ -18,6 +19,7 @@ import uk.gov.ons.ctp.integration.rhcucumber.selenium.pages.Pages;
       RespondentDataRepository.class,
       TestCloudDataStore.class,
       FirestoreDataStore.class,
+      FirestoreProviderImpl.class,
       Pages.class,
       FirestoreDataStore.class
     })
