+++ src/test/java/uk/gov/ons/ctp/integration/rhcucumber/repository/RespondentDataRepository.java
@@ -6,10 +6,10 @@ import javax.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import uk.gov.ons.ctp.common.cloud.TestCloudDataStore;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.common.event.model.CaseUpdate;
 import uk.gov.ons.ctp.common.event.model.UacUpdate;
+import uk.gov.ons.ctp.common.firestore.TestCloudDataStore;
 
 @Service
 public class RespondentDataRepository {
